Adds a fallback controller
The FallbackController is mapped to all paths that are not currently
handled by other controllers, this is necessary to allow the React
frontend app to handle the routing client-side.

Also adds some standard profiles that match environments and disallows
access to `/actuator` paths when the profile is `PROD` or `STAGING`.