Handles GH app installation webhook
When the GH app is installed in a org, we record the installation_id
and also handle subsequent modifications to the installation.

Also introduces an exception hierarchy to be used inside the console
app and a set of helper methods, in the Ensure class, to handle common
preconditions that we can to enforce in code.