Exchanges a JWT token for Session Keys in order to call StartBuild
The Session Keys are cached in memory for the user.
This also changes how users are created in that we call getId
for the user on creation and store their cognito identity id in the DB.
This change also moves us to the production cognito user pool and
identity pool.