Handles Push events to the GitHubApp.
When a PR is merged or a commit is pushed to main or master, our GitHub
app will receive a webhook event and trigger a CI build.

This PR also pushes the creation of the "SENT_TO_BUILD" build event
from ProjectApiController to AwsService.

One problem we have right now is that we don't have a way to map a
GitHub user to an Xtages user so it's unclear how to determine who
kicked off the build (should be the author of the commit if the build
is started from the webhook that the GH app receives). Until we figure
that one out, I'm using the org's owner as a stand-in.