Wires the project page to server data
Currently in HomePage we are showing the list of project templates
and the list of project the organization has.

This PR also adjusts xtages-internal-api.yaml so the includeLastBuild
is a query parameter (since we call GET /project) and to make some
Schema properties required (so we get better type checking).

Also adds a couple of layout components in Section and SectionTitle.