Implements the API to get a project details
GET /project/{name} will return the project details and if includeBuilds
is true then all builds will also be fetched.

I also refactored the response from /project so instead of returning
a ProjectAndLastBuild pojo I return a Project that contains a list of
Builds which might be empty, have the lastest Build or have all the
Builds for a Project.