Implements the usage section of the account page. (#107)
The Usage section of the account page is composed of UsageCards each
one for a ResourceType.

This PR also adds a better error message when trying to create a new
project and the user hits the limit.

Finally this PR introduces LoadIndicatingSection which is a Section that
is backed by data that is feched by useQuery. This component will
consistently apply a spinner (when the data is being loaded) and show
an error message if the data fails to load.