Adds a project settings API
The API currently only allows to associate an external domain name
to a Project. This will request a certificate from ACM and will return
the DNS record that the user should add to their DNS server (likely
at the registrar) to validate that the domain in fact belongs to them.