Implements the project settings page
The page currently only allows the user to associate a domain
to their project.

Once a domain is associated, we don't allow modifications (unless
the user contacts us). This is a simplificaton for time sake.

This PR also refactors the ProjectDetailsCard and extracts a
SimpleProjectCard that is the base of more complex project card
components.

In the backend we had to implement a retry strategy after
requestCertificate is called because AWS will return partially filled
CertificateDetails when calling describeCertificate immediately after
calling requestCertificate :(.
See https://github.com/aws/aws-sdk-js/issues/2133 and
https://github.com/hashicorp/terraform-provider-aws/blob/af2d5c0ad54f38f35898413300892b5a58f78291/aws/resource_aws_acm_certificate.go#L312