Updating the build information (#138)
* Updating the build information
This will update the deploy information when a deploy / undeploy happens

* Listen to Cloudtrail event for ECS

* Listening to UpdateService for deployment
When there is a deployment that is not first one that created the
service the ECS event `SERVICE_DEPLOYMENT_COMPLETED` is not triggered.
This adds a CloudTrail event to listen when the service is updated with
a desiredCount = 1. It's still not fullproof because we are listening to
when the service is redeploy but not when the service is in a steady
state. Still, it's better than nothing now.

* Quite some changes
- adding project deployment status
- fixing some warnings for Int64 in openAPI definition
- Extracting environment type
- Still pending the deploy status for the project in the Deployment
  object

* Updating queries and deployment object
This updates the deployment object to have the project deployment
status.
I have also updated the query for `fetchLatestDeploymentStatus` to
eliminate the subquery

* fixing query to get latest project deployment

* addressing review comments