Implements email notifications
Users are notified when a Build fails or when it's fixed (had previously
failed).

Emails are composed as [Rocker](https://github.com/fizzed/rocker) templates.

A new Gradle sub-project (email) was added, since it's necessary to
first compile rocker to java. Without the sub-project, the root project
compiles Kotlin first and then Java (meaning that the Kotlin code
wouldn't find the generated Java code because it hasn't been compiled
yet.)

The email templates are based on
[Cerberus](https://tedgoas.github.io/Cerberus/) with an Xtages
"theming".