Stops asking for name and org at sign-up.
Makes the `Principal`'s `name` and `org` properties `Nullable`.
Because we can't modify the Cognito User Pool to not require a `name`
attribute, we are using `__u_n_d_e_f_i_n_e_d__` as a marker value
to determine if the user has set their name.
Fixes the how we retrieve the priceId with a marker value.