Adds a Setup wizard
The wizard shows up on the home screen when the user has created
their account but hasn't installed the GH app and/or selected
a Plan.

The first step of the wizard is to create an account (always shows
up as completed). Second step is to install the GH app at which
point we will create the Xtages org. Third step is to select
and purchase a plan through Stripe.