Free tier (#212)
* Adds a Setup wizard

The wizard shows up on the home screen when the user has created
their account but hasn't installed the GH app and/or selected
a Plan.

The first step of the wizard is to create an account (always shows
up as completed). Second step is to install the GH app at which
point we will create the Xtages org. Third step is to select
and purchase a plan through Stripe.

* changes for free tier offering

* free tier provisioning

* checking that the org doesn't have a plan
- adding paid column in the plan

* checking for no plan as req for FT

* removing unused code

Co-authored-by: Carlos Zuniga <czuniga@xtages.com>