Refactors resources
Refactors our resources representation from the server. We were
previously conflating 3 concepts in our API: a type of resource, how a
resource is billed and an actual resource as provisioned. This PR cleans
up those concepts by renaming the `ResourceType` values so they only
represent generic resources, introducing `ResourceBillingModel` which
represents how a resource is billed and removing the `Resource` enum
which will be replaced by a Resource **object** in a follow-up PR.

This PR also consolidates the facilities to query resource usage under
the `/resource` path.